add default patience to all tests
- some refactoring
